Fix dylib install name for test imported XCFrameworks
Currently, generated XCFrameworks (and frameworks) binaries are declared
appending the Apple SDK identifier at the end of the file name. Binary filename
is used during dylib declaration, and sets the install name for the dynamic library
as:

`@rpath/MyFramework_iphonesimulator.framework/MyFramework_iphonesimulator`

Which is not expected and not the final Apple framework name. This change address
that issue and introduces an "intermediates" directory to clean up Bazel declared files.

PiperOrigin-RevId: 448593860