Fix import style in static xcframeworks
Currently the apple_static_xcframework rule produces a xcframework that
contains a static library and headers. The import style that was used
previously doesn't support that case, so they were not consumable via
Xcode, or our import rule. Now in the case that we use a non-framework
modulemap file, we also use quoted headers.