Allow no local provisioning profiles
On brand new, or CI machines, you might not have local profiles. You
also might not provide fallback profiles if you only want to use this
rule for picking up local profiles. Now both options are supported but
it errors in the case you have neither which is a better error for this
case than if the python script just failed.