Add support for UUID in profile finder (#1496)
Add the ability to use UUID (in addition to name) to identify provisioning profiles on the users system. This helps when you want 100% specificity around which profile to use. Use `expected_specifier` as the param name to match Xcode's `PROVISIONING_PROFILE_SPECIFIER` setting name.

This allows for both to be used:
```Starlark
    provisioning_profile = select({
        "//:Debug": "bb393eb3-d030-4796-8a5d-b67243fa1234",
        "//:AppStore": "AppStoreProfile_2022-06-01",
    }),
```