Refactor `_deduplicate` into `resources` module
By using a lambda, this becomes a reusable algorithm, with potential uses outside of its current sole use site. By moving it to the `resources` module it can be used from multiple locations, including outside of rules_apple, either with an internal-ignoring import, or if we make `resources` public in the future.