Move depset creation for ObjC provider fields to framework_import_support
Currently other fields for the ObjC provider creation for (xc)framework
imports receive a list instead of a depset object. This change moves
framework file and library fields to be list instead of depset.

Additionally this fixes an issue with current implementation for
`alwayslink` which fails because it's creating a depset of an object
that it's already a depset.

PiperOrigin-RevId: 456307958
(cherry picked from commit 8f84a4c92dfead8fdb45515f63e5a3bd40c09656)