Remove unused apple_xcframework attr.
This dates back to the early architecture of the API, which was based on prelim understandings of what Swift Package Manager could accept for XCFrameworks.

In practice it only accepts three types; (dynamic) frameworks, dynamic libraries (which are not useful outside of macOS or testing flows for [i|tv|watch]OS), and static libraries.

PiperOrigin-RevId: 457579889