Remove test for Swift consuming imported Swift static framework
Currently this test is working due to a regression introduced in
https://github.com/bazelbuild/rules_apple/commit/af389568de912f4e48f5baa50d8e77b4d99e718a, since the generated framework is using an Objective-C
header defining the same class and methods that `SharedClass.swift`
defines. However, once the generated framework uses the Swift
generated header, this will fail because swiftc will search for a
`.swiftmodule` file thanks to the `external_source_symbol` annotation
on the generated header that states the header language is Swift.

PiperOrigin-RevId: 458262562