Make testrunner attributes public
In rules_ios we're building out Apple Silicon based ephemeral VM test
runners to improve running in remote execution and locally. Currently,
we use rules_apple xctestrunner to run the tests under Bazel directly.
You can find this PR here:
https://github.com/bazel-ios/rules_ios/pull/509

Today if you'd like to make a different test runner, you have to copy
the rule and making the attributes public would help hook into them
better. The archicture as it stands is quite nice. Ideally we could make
`sim_creator` an implementation detail of the testrunner / template
combo: or feed that in a different way but leaving as-is for now