Fix race condition between xcodebuild and symlink action for test XCFramework
This change updates when modulemap files are generated XCFramework
with static libraries. Previously, modulemap were generated after
the XCFramework bundle was created with xcodebuild. Now, it's being
included as part of the headers directory.

Additionally, this change adopts the intermediates module ahead of a
rule migration to test/starlark_tests/.

PiperOrigin-RevId: 465638790