Raise version number to minimum of 12 or equivalent where possible on iOS/tvOS test targets for Starlark.
This is done to accomodate when arm64e was formally introduced for iOS and tvOS, and hopefully to address future test issues when building for this arch.

As armv7 will be dropped in Xcode 14 RC as a buildable architecture, we are using arm64e when possible to handle the representation of fat binaries for iOS devices.

PiperOrigin-RevId: 466809744