Standardize on minimum OS versions for Apple rules Starlark tests for easier maintenance.
This is largely based around the minimums defined for Xcode 14 RC and when the arm64e architecture was introduced for iOS and tvOS. However, a comprehensive listing of the different types of min OSes considered for our present day targets has been accounted for in common.bzl.

Relying on values as given by https://developer.apple.com/support/xcode/ and past Xcode release notes.

PiperOrigin-RevId: 468232906