Update iOS app tutorial (#1639)
This is a continuation of the work started in https://github.com/bazelbuild/rules_apple/pull/1621.

This PR contains a couple of improvements to the iOS tutorial:
- I refreshed the dated wording and instructions of the overall tutorial (bumped the rules version, `minimum_os_version`, etc.).
- Added complete instructions to generate an `xcodeproj` using [`rules_xcodeproj`](https://github.com/buildbuddy-io/rules_xcodeproj). If we prefer to also link to Tulsi, let me know and we could simply link to this [new document in that repo](https://github.com/bazelbuild/tulsi/pull/354).
- Added a mention of `bazel run` to directly run an iOS app in the iOS Simulator, along with some useful flags to customize the device and OS version used.
- Added a mention of some new rules to handle provisioning profiles, which could be useful to newcomers.

This gets the tutorial into a more modern state. I will need to make a PR to the https://github.com/bazelbuild/examples soon to update the final state of the project to include the updates made here. We could also decide to rewrite the sample app in Swift instead of Objective-C, but not sure how important that is at the moment.

Co-authored-by: JP Simard <jp@jpsim.com>
Co-authored-by: thii <t@thi.im>
Co-authored-by: Brentley Jones <github@brentleyjones.com>