Lazy hash compute in bundletool.py
To prevent multiple files with different content be written into the same dest,
bundletool.py compare files by their hash value before writing. But in most
cases, there are no conflicting files, and the time used to calculate the hash
is wasted. This commit defers the calculation of the hash until there are
actually two files that need to be written to the same dest.