[rules_apple] Migrate static_framework_file to imported_library
STATIC_FRAMEWORK_FILE is a static archive, and we should be able to
migrate it to IMPORTED_LIBRARY. We will eventually migrate it to a
LibraryToLink in the CcLinkingContext.

If this migration breaks build, it is likely because there are two
~equivalent static frameworks in the dependency, that were
deduplicated when -F are used to find them. We should fix those builds
by eliminating the redundant dependency.

PiperOrigin-RevId: 472471843