Remove repo name from module import (#1695)
This is in preparation for adding support for bzlmod. This Python script
loads another file via its runfiles. When bzlmod is enabled, the
directory has the name of the actual dependency and its version and not
the name of its defined `repo_name` (`rules_apple~1.1.2` instead of
`build_bazel_rules_apple`). This causes issues and should be addressed
in Bazel (which it is being thought about per [this Slack
conversation](https://bazelbuild.slack.com/archives/C014RARENH0/p1664279459175029)),
but this is a quick fix that should be backwards compatible.