Adding ownership for apple_resource_group and apple_resource_bundle (#1745)
This PR adds a default owner for when
[`resources.dedeplicate`](https://github.com/bazelbuild/rules_apple/blob/master/apple/internal/partials/resources.bzl#L268-L272)
is called. When an apple_framework target (from rules_ios) is built as
dynamic, the resources_partial in rules_apple is called. This leads to
`resources.deduplicate` being called with avoid_providers populated with
`AppleResourceInfo` providers from the `apple_resource_group` or
`apple_resource_bundle` target. This calls [_merge_proiders with
validate_all_resources_owned =
True](https://github.com/bazelbuild/rules_apple/blob/master/apple/internal/resources.bzl#L888-L891),
which results in `Error in fail: The given providers have a resource
that doesn't have an owner, and validate_all_resources_owned was set.
This is most likely a bug in rules_apple, please file a bug with
reproduction steps.`

This PR adds a default owner to that call so the the
validate_all_resources_owned check will pass.