Add the `apple.no_legacy_swiftinterface` feature to disable the automatic library evolution transition on framework rules.
We are migrating users to a model where a new attribute, `library_evolution`, must be set on `swift_library` targets that vend the user-facing module in a framework, instead of automatically applying library evolution to the whole subgraph. In order to migrate, users can add this feature to their framework target and at the same time set `library_evolution = True` on their library.

Once everyone is migrated, library evolution should be removed entirely from the Apple transitions and the feature can be removed from BUILD files.

PiperOrigin-RevId: 487022298