Ignore skip_codesign_simulator_bundles for frameworks when building for device (#1770)
We recently enabled `--features=apple.skip_codesign_simulator_bundles`
and we just noticed that builds for the device now fail to install with
the error: `No code signature found.`

We have a dynamic framework that is included in the app which is shared
with extensions, and running `codesign --verify` on the framework gives
this:
```
MyFramework.framework: code object is not signed at all
```
If we disable `--features=apple.skip_codesign_simulator_bundles` we get
this instead:
```
MyFramework.framework: valid on disk
MyFramework.framework: satisfies its Designated Requirement
```

Going through the code, it seems like there's a missing case for
checking if we're building for the device, we should still sign
frameworks. As you can see at line 450, we were checking for that
already, and now I added the same check at line 430. I tested it in our
project and it works, but let me know if you think this should be fixed
in some other way.