Add xctestrun based iOS unit test runner (#1612)
This test runner is a drop in replacement for the default iOS test
runner that you can use by setting it on the `runner` attribute of an
`ios_unit_test`. It uses the `xctest` CLI directly for logic tests, and
uses `xcodebuild test-without-building` with a `xctestrun` file for
hosted tests. It supports running tests in random order, and generating
`xcresult` bundles from any type of test. It also supports running
x86_64 test bundles on arm64 hosts. Examples can be found in the docs.