Add bzlmod module_extension for provisioning_profile_repository
This is required to be able to use provisioning_profile_repository at
the same time as using rules_apple with bzlmod. The setup is a bit
different but the gist of it is that we run the extension by default
(meaning users don't have to run it at all unless they need to specify
fallback_profiles), and then if the user runs it as well we fetch the
data from their run instead.