Fix duplicate symbols for static frameworks (#1859)
As part of moving the linking info into CcInfo (as well as objc info
during the migration). Static frameworks started cause duplicate symbols
in some cases. This was because bazel seems to dedup `library` between
the 2 providers but not `static_framework_file` since
`static_framework_file` has been removed from google's use. Otherwise
this
change seems to be a no-op, probably because we're bazel 6.x+