Make emitswiftinterface parameter public
On repositories where ios_static_framework() are built along side with
other ios_*() targets, the emit-swiftinterface transition duplicates all
the actions downstream of that transition.
This is by design, but it may not be desirable in all cases. For
instance, if the swift code depends on a lot of cc_*() and objc_*()
targets, it may be preferrable to not transition and add the
swift.emit_swiftinterface feature to swift modules manually.

This changes makes the emitswiftinterface option to
ios_static_framework() public so that the caller can opt-in to disable
this transition and manage this feature manually.