Change bundle_id_suffix default for ios_imessage_application containers to be the equivalent of an empty string.
The nature of the bundle ID for an iMessage application ( a self-hosting app container ) is that its bundle ID needs to be the prefix of all hosted extensions ( iMessage extensions, Sticker Pack extensions ). As such, its handling of the composed bundle ID should resemble that of ios_application rather than ios_extension.

PiperOrigin-RevId: 534463715