Force _main symbol for application extensions
In the case you're linking with LTO, the `_main` does not get loaded
correctly because it seems that `-Objc` doesn't force static libraries
to be loaded in the case the objects in the archive are bitcode. This
makes sure the main file is loaded and that can cascade to the rest of
the used code. This is only needed because `-e _NSExtensionMain` makes
the linker not need to load the standard `_main`.

https://github.com/apple/swift/issues/67460