Inline executable_name reading (#2074)
In a future cherry pick we will no longer be able to get the bundle_name
with just the ctx argument as we do for the executable_name fallback.
Instead of passing along all the new required arguments this change
inlines the executable_name attr reading and passes that along, and
makes the consumers of that arg do the behavior of
`executable_name or bundle_name` themselves.