Add support for watchOS single target applications (#2048)
This PR cherry-picks 4 different commits that add support for a first
version of a watchOS single target applications that was introduced in
Xcode 14. There are more commits in the upstream branch that refactor
the rules later on, so I will do more cherry-picks before these rules
should be publicly used, but this PR is already pretty big as it is.