Include Apple Symbols for watchOS extensions (#1942)
This PR depends on https://github.com/bazelbuild/rules_apple/pull/1941
since I'd like to fully verify the tests before merging this.

This fixes bundling Apple Symbols for watchOS extensions correctly when
`include_symbols_in_bundle` is set to True in the companion iOS app. The
problem seemed to be that `watchos_application` was missing a
`partials.apple_symbols_file_partial` implementation to propagate the
symbols to the parent iOS app. This caused all symbols for watchOS
extensions to be correctly created, but not included in the final
`Symbols/` directory.