Disable swiftinterface compilation support (#2163)
This current behavior fails in the real world and requires some changes
that are potentially unclear. I think if we have a reason to enable it
we can try to push those. We might also need a way for things to disable
this support since some frameworks like Firebase seem to have an invalid
layout for this, likely because they've never needed to make that work
before.