Only use label based archive roots if needed
This changes the logic that outputs bundles to label-based directories
to only apply if the label name and the bundle name differ, which is the
best approximation for "whether or not this is needed" without
introducing a switch for it. The only downside is the "primary" target
that has the matching label and bundle name uses a different output
directory structure from the other ones. At least in the case of using
separate labels for debug vs alpha etc all the targets probably differ.