Add validation for platform type attribute in `apple_{,static}_xcframework` rule
The platform type attributes (`ios`, `tvos`,...) aren't mandatory, but
at least one of them has to be set. It wasn't clear in the documentation
that one of these is required, because there is another attribute
`minimum_os_versions` which also mentions platform types.

Previously not setting any of these led to an error in:

```
Error in link_multi_arch_static_library: unexpected empty key in split transition
```

which wasn't that much useful.