[Docs] Refresh iOS tutorial (#2314)
[Recently](https://github.com/bazelbuild/examples/pull/331), the
`ios-app` project living in
[bazelbuild/examples](https://github.com/bazelbuild/examples) was
removed. This PR contains a few improvements:
- I updated our tutorial that shows how to create create a simple iOS
app that builds with Bazel not to rely or point to the non-existing
project.
- We now provide all the source code (one Swift file and one
`Info.plist`) to generate a basic sample iOS app in this tutorial.
- The source code is now Swift instead of Objective-C.
- Simplified many sentences and commands, along with a few bug fixes
required due to breaking changes in `rules_xcodeproj` (it now requires a
`@bazel_features` explicit dependency).

I think it would be a good idea to move the tutorial to use bzlmod once
Bazel 7 ships. For now, we can continue using the `WORKSPACE` for ease
of compatibility with the default Bazel 6 behavior.