Remove subset of valid values from `families` attribute documentation (#2326)
This is a bit of a documentation nitpick. Rules can have different
supported families. The documentation for all rules was hardcoded to
mention that valid values are only `iphone` and `ipad` which isn't
correct when looking at `watchos_*`, `tvos_*`, `visionos_*`, and
`macos_*` rules.

I simply vote for removing the list of valid values from the
documentation for `_device_family_attrs` given it is shared by all
rules. We could specifically pass into this attribute the type of rule
so that we can adapt the documentation for each platform to mention the
allowed values, but I don't think it's really worth the complexity. Most
rules only support one family (everything except iOS from my
understanding), so people shouldn't really need to touch this attribute.