Add swiftinterface_imports to apple_static_xcframework_import (#2328)
Without this static xcframeworks which have a swiftinterface file will
fail to compile when sandboxing is enabled since their swiftinterface
won't be copied to the xcframework bundle.

This changed in https://github.com/bazelbuild/rules_apple/pull/2266
which was mirroring the swiftmodule_imports behavior that I believe is
also wrong so I fixed that here too.