Fix private header support in `experimental_mixed_language_library` (#2422)
I was running into an issue where an
`experimental_mixed_language_library` declared a private header via
`srcs`. A type from this header was used in the Swift module part of the
mixed library leading to not found errors.

This adds the private headers to the module map we create for the
`swift_library` allowing it to find the types in the those headers. Note
that because this is not the "umbrella" modulemap, the private headers
do not end up being exposed publicly to dependencies of this mixed
library and are only available within the mixed library as intended.