Add apple_metal_library rule (#2421)
This rule should be used in situations where the default resource
processing behavior is insufficient, such as when custom copts are
required. It is an adaptation of the `metal_library` rule found in
[rules_apple_line](https://github.com/line/rules_apple_line/blob/main/apple/metal_library.bzl),
with changes made to more closely resemble `apple_intent_library`.

Here's an example:

```starlark
apple_metal_library(
    name = "metal_lib",
    copts = ["-w"],
    hdrs = ["shared.h"],
    srcs = ["shader1.metal", "shader2.metal"],
)

objc_library(
    name = "my_lib",
    ...
    data = [":metal_lib"],
)
```

This PR should be additive to how Metal is supported as data today, and
should not introduce any breaking changes.