Use buildifier 6.4.0 for improved load sorting (#2500)
Followup to/depends on #2499 

Bigger PR than 2499 due to me not including `.bzl` files. I was using a
command from my history that just operated on BUILD/MODULE files. Big
enough it might make upstream merges a little painful? I'll leave that
thought to the maintainers.

**6.3.3:** `out_of_order_load` warning is enabled by default
**6.4.0:** More correct sorting order of load statements

Could bump to 7.1.2 which is latest, but not necessary for auto fixing
of load sorting