Support Linux with `dossier_codesigning_reader.py`
On Linux there is no keychain or the `security` command. To codesign on Linux you’ll have to use something like https://gregoryszorc.com/docs/apple-codesign/0.17.0/apple_codesign_rcodesign.html, with a wrapper that translates `codesign` arguments to `rcodesign` arguments.

Using that method requires using `.cer` files instead of identity strings (which are part of the subject of a certificate). The change made here assumes you are using that method, which has worked for us in our internal project.

Signed-off-by: Brentley Jones <github@brentleyjones.com>