Remove analysis failures for use_tree_artifacts_outputs flag.
We started using `apple_*_xcframework_import` rules to bring in XCFrameworks and we are getting failures during a query.
We don't actually use the version macOS framework in our build, it's just shipped along side the iOS framework.

Because of the `fail` we need to remove usage of the `use_tree_artifacts_outputs` flag, which causes a big regression to our disk usage:

- bin/Code/Apps/CashApp/CashApp_archive-root/Payload/Cash.app (732 MB)
- bin/Code/Apps/CashApp/CashApp-intermediates/unprocessed_archive.zip (732 MB)
- bin/Code/Apps/CashApp/CashApp.ipa (740 MB)

With `use_tree_artifacts_outputs` flag enabled we only have:

- bin/Code/Apps/CashApp/CashApp_archive-root/Payload/Cash.app (732 MB)

We should see about better supporting this flag in the future, consider making it the default, or improve the disk usage without it.