Improve `MainThreadCheckerLibsCopy` performance
The `MainThreadCheckerLibsCopy` action was showing up as a slow action in our perf. dashboard.
This action simply copies the dylib from Xcode into the tests runfiles but was taking over 150ms on average to run.

This action used Python to run and given it's so simple this PR replaces it with a `.sh` script `sh_binary` which is much simpler and faster.

With this change over 3 runs:

- Before: `164.6 ms` average
- After: `18.2 ms` average