Add support for `.provisionprofile` in `local_provisioning_profile`
macOS profiles have the `.provisionprofile` extension so `local_provisioning_profile` would always fail to find any macOS provisioning profiles. I added a new `profile_extension` attr to the rule to support specifying the profile is for macOS as it's the only platform that seems to use this extension I kept the `.mobileprovision` extension as the default.