Support pre- and post- actions in `ios_xctestrun_runner` (#2649)
Iterating on this out in the open. It's a proof of concept, but it does
work. I'm not married to any aspect of this API, the fact that hooks
have access to all environment variables the test runner does, nor the
lack of a live log-parser hook in this implementation. I recommend
maintainers be ruthless and specific in what they want out of this
change.