Add `exec_data` attribute to tests
This differs from `data` in that it uses the `exec` config for the files. This is useful for running host-based tools in UI tests.

We could add `data` to the test runners for the same effect, but then we would need different test runners per test, which isn’t ideal.

Signed-off-by: Brentley Jones <github@brentleyjones.com>