wip: run presubmit CI in github actions
experimenting to potentially move off of bazelci, due to various challenges. this is a 1:1 representation of the presubmit.yml as-is, meaning buildifier hasn't been updated from 6.4.0, xcode is still 16.2, and pins for doc_tests are still in place. updating buildifier and moving the stardoc/doc_tests pins should be done in a separate PR, and potentially xcode 16.3 as well (this would also allow us to more easily test a matrix of xcodes, if desired).