Use `TEST_TMPDIR` where possible (#2686)
`TEST_TMPDIR` is in the Bazel exec root. This is better than `/tmp` or
`$TMPDIR` since it’s managed by Bazel and will be cleaned up with `bazel
clean`. Also, when using remote execution, it allows the output files to
stay within the defined output tree which the executor can clean up
properly.

Signed-off-by: Brentley Jones <github@brentleyjones.com>