Update rules_apple to use the the new `swift_common` toolchain APIs.
Also remove the explicit execution group and the workaround that disabled auto-exec-groups, because the underlying problem with the Swift rules has been fixed.

Note: Even though `swift_toolchain.use_all_toolchains()` also requests the C++ toolchain, adding `use_cpp_toolchain()` to that is fine; the duplication is harmless. This lets us express more explicitly "the rule depends on the Swift (and C++) toolchains, but it also uses the C++ toolchain for other things".
PiperOrigin-RevId: 755951150