Ability to explicitly specify output files in apple_core_data_model rule (#2727)
This PR contains one of the two fixes mentioned in #2726. This change
makes it possible to explicitly specify the output files that are
expected to be generated from the provided core data models. This causes
the rule to return a provider containing these individual Swift files
rather than just the containing folder.