transform XCTRunner.app Info.plist to xml before sed'ing it (#2763)
Handle the new case in Xcode 26 where the Info.plist in the
XCTRunner.app is a binary plist, rather than an XML plist in Xcode 16.
Coercing the plist to xml works whether or not it is originally a plist,
and Xcode 16 handles the case of a binary plist adequately.

Also willing to rework this to do keypath-based substitution instead of
using sed, though the sed approach has definitely proven resilient up to
now.