environment_plist: lower the entire --platform string
The `xcrun --sdk` argument expects only lowercase letters and numbers.
SDK version is usually just numbers, but in the exceptional case where
it also has letters, we need to .lower() the entire string that we pass,
and not just the platform name. We're slightly modifying
`_environment_plist_impl()` in that commit to support that use-case.